+++ visidata/keys.py
@@ -6,6 +6,7 @@ visidata.vd.prettykeys_trdict = {
         ' ': 'Space',  # must be first
         '^[': 'Alt+',
         '^J': 'Enter',
+        '^M': 'Enter',
         '^I': 'Tab',
         'KEY_UP':    'Up',
         'KEY_DOWN':  'Down',
@@ -16,8 +17,6 @@ visidata.vd.prettykeys_trdict = {
         'KEY_EOL':   'End',
         'KEY_PPAGE': 'PgUp',
         'KEY_NPAGE': 'PgDn',
-        'KEY_IC':    'Ins',
-        'KEY_DC':    'Del',
 
         'kUP5':      'Ctrl+Up',
         'kDN5':      'Ctrl+Down',
@@ -30,10 +29,13 @@ visidata.vd.prettykeys_trdict = {
         'KEY_IC5':   'Ctrl+Ins',
         'KEY_DC5':   'Ctrl+Del',
 
+        'KEY_IC':    'Ins',
+        'KEY_DC':    'Del',
+
+        'KEY_SRIGHT':'Shift+Right',
         'KEY_SR':    'Shift+Up',
         'KEY_SF':    'Shift+Down',
         'KEY_SLEFT': 'Shift+Left',
-        'KEY_SRIGHT':'Shift+Right',
         'KEY_SHOME': 'Shift+Home',
         'KEY_SEND':  'Shift+End',
         'KEY_SPREVIOUS': 'Shift+PgUp',