+++ visidata/loaders/pandas_freqtbl.py
@@ -166,12 +166,12 @@ class PandasFreqTableSheet(PivotSheet):
                 {}
             ))
 
-def expand_source_rows(source, vd, cursorRow):
+@Sheet.api
+def expand_source_rows(sheet, row):
     """Support for expanding a row of frequency table to underlying rows"""
-    if cursorRow.sourcerows is None:
-        vd.error("no source rows")
-    vs = PandasSheet(source.name, valueNames(cursorRow.discrete_keys, cursorRow.numeric_key), source=cursorRow.sourcerows)
-    vd.push(vs)
+    if row.sourcerows is None:
+        vd.fail("no source rows")
+    return PandasSheet(sheet.name, vd.valueNames(row.discrete_keys, row.numeric_key), source=row.sourcerows)
 
 PandasSheet.addCommand('F', 'freq-col', 'vd.push(PandasFreqTableSheet(sheet, cursorCol))', 'open Frequency Table grouped on current column, with aggregations of other columns')
 PandasSheet.addCommand('gF', 'freq-keys', 'vd.push(PandasFreqTableSheet(sheet, *keyCols))', 'open Frequency Table grouped by all key columns on source sheet, with aggregations of other columns')
@@ -179,6 +179,10 @@ PandasSheet.addCommand('gF', 'freq-keys', 'vd.push(PandasFreqTableSheet(sheet, *
 PandasFreqTableSheet.addCommand('t', 'stoggle-row', 'toggle([cursorRow]); cursorDown(1)', 'toggle selection of rows grouped in current row in source sheet')
 PandasFreqTableSheet.addCommand('s', 'select-row', 'select([cursorRow]); cursorDown(1)', 'select rows grouped in current row in source sheet')
 PandasFreqTableSheet.addCommand('u', 'unselect-row', 'unselect([cursorRow]); cursorDown(1)', 'unselect rows grouped in current row in source sheet')
-PandasFreqTableSheet.addCommand(ENTER, 'open-row', 'expand_source_rows(source, vd, cursorRow)', 'open copy of source sheet with rows that are grouped in current row')
+PandasFreqTableSheet.addCommand(ENTER, 'open-row', 'vd.push(source.expand_source_rows(cursorRow))', 'open copy of source sheet with rows that are grouped in current row')
 
 PandasFreqTableSheet.class_options.numeric_binning = False
+
+vd.addGlobals({
+        'PandasFreqTableSheet': PandasFreqTableSheet,
+        })