+++ visidata/statusbar.py
@@ -80,8 +80,6 @@ def debug(vd, *args, **kwargs):
     'Display *args* on status if options.debug is set.'
     if options.debug:
         return vd.status(*args, **kwargs)
-    else:
-        return vd.addToStatusHistory(*args, **kwargs)
 
 def middleTruncate(s, w):
     if len(s) <= w:
@@ -108,9 +106,9 @@ def drawLeftStatus(vd, scr, vs):
     cattr = colors.get_color('color_status')
     active = (vs is vd.activeSheet)
     if active:
-        cattr = update_attr(cattr, colors.color_active_status, 0)
+        cattr = update_attr(cattr, colors.color_active_status, 1)
     else:
-        cattr = update_attr(cattr, colors.color_inactive_status, 0)
+        cattr = update_attr(cattr, colors.color_inactive_status, 1)
 
     if scr is vd.winTop:
         cattr = update_attr(cattr, colors.color_top_status, 1)
@@ -178,7 +176,7 @@ def drawRightStatus(vd, scr, vs):
     active = vs is vd.activeSheet
 
     if active:
-        statcolors.append((vd.prettykeys(vd.keystrokes) or '', 'color_keystrokes'))
+        statcolors.append((f'{vd.prettykeys(vd.keystrokes)} ' or '', 'color_keystrokes'))
 
     if vs.currentThreads:
         statcolors.insert(0, vd.checkMemoryUsage())
@@ -197,9 +195,9 @@ def drawRightStatus(vd, scr, vs):
                 if scr is vd.winTop:
                     cattr = update_attr(cattr, colors.color_top_status, 0)
                 if active:
-                    cattr = update_attr(cattr, colors.color_active_status, 0)
+                    cattr = update_attr(cattr, colors.color_active_status, 1)
                 else:
-                    cattr = update_attr(cattr, colors.color_inactive_status, 0)
+                    cattr = update_attr(cattr, colors.color_inactive_status, 1)
                 statuslen = clipdraw(scr, vs.windowHeight-1, rightx, rstatus, cattr.attr, w=vs.windowWidth-1, rtl=True)
                 rightx -= statuslen
                 ret += statuslen