+++ visidata/textsheet.py
@@ -7,7 +7,7 @@ import visidata
 
 
 vd.option('wrap', False, 'wrap text to fit window width on TextSheet')
-vd.option('save_filetype', '', 'specify default file type to save as', replay=True)
+vd.option('save_filetype', 'tsv', 'specify default file type to save as', replay=True)
 
 
 ## text viewer